Orchestrator

The Orchestrator can be thought of as a locally running daemon that handles running the various Searchers. There are three main commands that are useful to interact with it (from a standard shell):

romeo orchestrator start: starts the Orchestrator and loads and starts the searchers that were previously running

romeo orchestrator stop: stops and saves the currently loaded Searchers and gracefully stops the Orchestrator

romeo orchestrator status: shows an overview of the currently loaded Searchers

For more information use -h after these commands in the shell.

Advanced

RemoteREPL

Behind the scenes the Orchestrator runs in a RemoteREPL, which can be accessed to interact directly with the loaded Searchers. This can be achieved by starting a julia REPL and executing the following commands:

using RomeoDFT
connect_orchestrator()

You will see that now by pressing > you get a pink julia prompt, which means you're executing commands inside the Orchestrator REPL. The useful commands here are:

start_searcher("<searcher_name>"; verbosity = <number>): starts a Searcher with the provided name with a given logging verbosity (higher = more logged).

stop_searcher("<searcher_name>"): stops a Searcher with the name

load_searcher: simply loads a Searcher from disk if no function is provided as the first argument. See load_searcher documentation for more info.

Submitting Searcher

In order to create and submit a Searcher manually you can use setup_search and orchestrator_submit in sequence.

Reference

RomeoDFT.start_searcher — Function

start_searcher(name; verbosity=0)

Starts a Searcher and sets the logging verbosity. Higher verbosity means more will be printed to the log file.

RomeoDFT.stop_searcher — Function

stop_searcher(name)

Stops a Searcher.

RomeoDFT.load_searcher — Function

load_searcher([f::Function,] name)

Loads a Searcher to the Orchestrator, and executes f on it if it is provided. This can be useful to perform manual tasks or retrieve more granular information about the Searcher.

Example

load_searcher("MnO") do s
    # execute code on Searcher s which represents the MnO searcher
end

Missing docstring.

Missing docstring for connect_orchestrator. Check Documenter's build log for details.

RomeoDFT.orchestrator_submit — Function

orchestrator_submit(searcher)

Save and submit searcher to the running Orchestrator.